Multiple: Added web authentication for extended APIs
This commit adds three main things: two php scripts that get web API auth
tokens, Python code to utilize the auth tokens (including code that gets
whether or not the song is saved), and Arduino code to display the save state
of the song. The Arduino code also includes a fix for the LCD display, as
explained in README.md, for the ST7565 library.