Arduino: Added scroll indicators
This commit adds arrow icons below each line to indicate the scroll status of
the line above. For example, a string that is too long for one screen will
have a right-pointing arrow below it. Once the end of the string is reached, a
left-pointing arrow is displayed. If there are screens before and after the
current one, both left- and right-pointing arrows are displyayed