Arduino: Modified SRAM allocation
Previously, the display strings were dynamically allocated. This was causing
problems with them having garbage data in them when very long data was stored
in them. This fix reserves the strings at the beginning of the sketch,
preventing fragmentation of the memory. Furthermore, literal strings were
enclosed in the F() macro in order to free up a little more RAM for usage
elsewhere. The default STRING_BUFFER_SIZE was also lowered, as the clock
requires more RAM.