Arduino & Python: Started two-way communication
Previously data would only flow from the computer to the Arduino. This commit
adds the capability for the Arduino to send configuration data to the Python
script. In this commit, the a string buffer size message was added, allowing
the Python script to know the value of STRING_BUFFER_SIZE in the sketch. This
will be used in a future commit.